server/requirements.txt
aiohttp==3.8.6
aioice==0.9.0
aiortc==1.5.0
aiosignal==1.3.1
numpy==1.23.5
opencv-python==4.8.1.78
pytest-asyncio==0.21.1
python-pytuntap==1.0.5
